requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
argcomplete==3.5.0
asgiref==3.8.1
asyncer==0.0.7
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chainlit==1.1.404
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.1
dependency-injector==4.42.0
Deprecated==1.2.14
deprecation==2.1.0
docx2txt==0.8
EbookLib==0.18
et-xmlfile==1.1.0
fastapi==0.110.3
filelock==3.16.0
filetype==1.2.0
flatbuffers==24.3.25
flatdict==4.0.1
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud==0.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langsmith==0.1.120
Lazify==0.4.0
literalai==0.0.607
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.22.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.0
openpyxl==3.1.5
oracledb==2.4.1
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pillow==10.4.0
posthog==3.6.5
proto-plus==1.24.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyodbc==5.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-pptx==1.0.2
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
setuptools==75.0.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
splunk-sdk==2.0.2
SQLAlchemy==2.0.34
sqlparse==0.5.1
sqlvalidator==0.0.20
starlette==0.37.2
sympy==1.13.2
syncer==2.0.3
tenacity==8.5.0
textract==1.5.0
oracledb
tokenizers==0.20.0
tomark==0.1.4
tomli==2.0.1
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
uptrace==1.26.0
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
xmltodict==0.13.0
yarl==1.11.1
zipp==3.20.2
pymongo==4.8.0
poetry==1.8.3
pre-commit==3.8.0
codespell==2.3.0
flake8==7.1.1
pep8-naming==0.14.1
uvloop==0.20.0; sys_platform != 'win32'
Jinja2==3.1.4
docling==2.39.0
mariadb==1.1.12
rapidocr_onnxruntime==1.4.4
langdetect==1.0.9
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
tiktoken==0.9.0
PyMuPDF==1.26.3